Remove base64 from groups too
